# queue_get_next() item retrieval performance upgrade