--- a/qcfractal/storage_sockets/models/sql_models.py
+++ b/qcfractal/storage_sockets/models/sql_models.py
@@ -251,7 +251,7 @@ class TaskQueueORM(Base):
         Index("ix_task_queue_created_on", "created_on"),
         Index("ix_task_queue_keys", "status", "program", "procedure", "tag"),
         Index("ix_task_queue_manager", "manager"),
-        Index("ix_task_queue_base_result_id", "base_result_id")
+        Index("ix_task_queue_base_result_id", "base_result_id"),
     )