--- a/qcfractal/tests/test_managers.py
+++ b/qcfractal/tests/test_managers.py
@@ -110,9 +110,10 @@ def test_queue_manager_multiple_tags(compute_adapter_fixture):
     }
     for result in ret:
         assert result.status == ref_status[result.id]
-
     manager.await_results()
     ret = client.query_results(tasks)
+    for result in ret:
+        print (f"here you go: {(result.id, result.status)}")
     ref_status = {
         tasks[0]: "COMPLETE",
         tasks[1]: "COMPLETE",